+++ src/components/home/Meal/MealList/style.ts
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const MealListContainer = styled.div`
+  width: 460px;
+  height: 231px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+`;
