+++ src/components/home/Meal/index.tsx
@@ -1,70 +1,19 @@
-import { EMealType } from "../../../enum/meal/meal.enum";
-import useMeal from "../../../hooks/meal/useMeal";
 import MealDatePicker from "./MealDatePicker";
-import MealItem from "./MealItem";
-import { MealContainer, MealItemWrap } from "./style";
-import MealBreakfastIcon from "../../../assets/icons/meal/morning.png";
-import MealLunchIcon from "../../../assets/icons/meal/afternoon.png";
-import MealDinnerIcon from "../../../assets/icons/meal/night.png";
-import dayjs from "dayjs";
-import dateTransform from "../../../util/transform/dateTransform";
-import isBetween from "dayjs/plugin/isBetween";
-dayjs.extend(isBetween);
+import { MealContainer } from "./style";
+
+import ErrorBoundary from "../../../components/common/ErrorBoundary";
+import { Suspense } from "react";
+import MealList from "./MealList";
 
 const Meal = () => {
-  const { date, meal, handleMealDate, prevMealDate, nextMealDate } = useMeal();
-
-  const { BREAKFAST, LUNCH, DINNER } = EMealType;
-  const currentTime = dateTransform.fullDate().split(" ")[1];
-  const currentDate = dayjs(dateTransform.fullDate(date + currentTime));
-
   return (
     <MealContainer>
-      <MealDatePicker
-        date={date}
-        handleMealDate={handleMealDate}
-        prevMealDate={prevMealDate}
-        nextMealDate={nextMealDate}
-      />
-
-      <MealItemWrap>
-        <MealItem
-          mealData={meal?.breakfast!}
-          mealType={BREAKFAST}
-          mealIconSrc={MealBreakfastIcon}
-          isMealTime={
-            currentDate.isBetween(
-              `${dateTransform.hyphen()} 05:00`,
-              `${dateTransform.hyphen()} 07:50`,
-              "minute"
-            ) && meal?.breakfast! !== null
-          }
-        />
-        <MealItem
-          mealData={meal?.lunch!}
-          mealType={LUNCH}
-          mealIconSrc={MealLunchIcon}
-          isMealTime={
-            currentDate.isBetween(
-              `${dateTransform.hyphen()} 07:50`,
-              `${dateTransform.hyphen()} 13:20`,
-              "minute"
-            ) && meal?.lunch! !== null
-          }
-        />
-        <MealItem
-          mealData={meal?.dinner!}
-          mealType={DINNER}
-          mealIconSrc={MealDinnerIcon}
-          isMealTime={
-            currentDate.isBetween(
-              `${dateTransform.hyphen()} 13:20`,
-              `${dateTransform.hyphen()} 19:10`,
-              "minute"
-            ) && meal?.dinner! !== null
-          }
-        />
-      </MealItemWrap>
+      <MealDatePicker />
+      <ErrorBoundary fallback={<>에러발생</>}>
+        <Suspense fallback={<>로딩중...</>}>
+          <MealList />
+        </Suspense>
+      </ErrorBoundary>
     </MealContainer>
   );
 };
