+++ src/store/meal/mealStore.ts
@@ -1,0 +1,7 @@
+import { atom } from "recoil";
+import dateTransform from "../../util/transform/dateTransform";
+
+export const mealDateAtom = atom({
+  key: "meal/mealDateAtom",
+  default: dateTransform.hyphen(),
+});
